make the page message tool modular
