+++ common/page-messages.js
@@ -1,0 +1,7 @@
+import bom60 from '../public/bom60.png';
+
+const messages = [
+  {header: "Book of Mormon in 60 Days",content:"The youth of our ward would like to invite you to join in on their Book of Mormon 60 day challenge for June and July! Many of them also committed to taking a book of Mormon, writing their testimony in it and giving it to a friend to join them in this challenge. They want to extend this invitation to you as well! The sister missionaries have a whole box of Book of Mormons! Lets all do out part to gather Israel!", image: bom60}
+]
+
+export default messages;
