+++ next.config.js
@@ -1,11 +1,6 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
-  experimental: {
-    images: {
-      layoutRaw: true,
-    },
-  },
 }
 
 module.exports = nextConfig
